fix: complete merging
