+++ lib/features/stories/view/widget/expanded_stories_section.dart
@@ -21,7 +21,6 @@ class ExpandedStoriesSection extends ConsumerWidget {
           mainAxisAlignment: MainAxisAlignment.start,
           children: [
             SizedBox(
-              height: kIsWeb ? kToolbarHeight : kToolbarHeight + 25,
               height: kIsWeb ? kToolbarHeight : kToolbarHeight + 26,
             ),
             Padding(
